+++ Gw2Sharp/WebApi/V2/Gw2WebApiV2Client.cs
@@ -56,6 +56,7 @@ namespace Gw2Sharp.WebApi.V2
         private readonly ISkinsClient skins;
         private readonly ISpecializationsClient specializations;
         private readonly IStoriesClient stories;
+        private readonly ITitlesClient titles;
         private readonly ITokenInfoClient tokenInfo;
         private readonly IWorldBossesClient worldBosses;
 
@@ -116,6 +117,7 @@ namespace Gw2Sharp.WebApi.V2
             this.skins = new SkinsClient(connection, gw2Client);
             this.specializations = new SpecializationsClient(connection, gw2Client);
             this.stories = new StoriesClient(connection, gw2Client);
+            this.titles = new TitlesClient(connection, gw2Client);
             this.tokenInfo = new TokenInfoClient(connection, gw2Client);
             this.worldBosses = new WorldBossesClient(connection, gw2Client);
         }
@@ -248,6 +250,9 @@ namespace Gw2Sharp.WebApi.V2
 
         /// <inheritdoc />
         public virtual IStoriesClient Stories => this.stories;
+
+        /// <inheritdoc />
+        public virtual ITitlesClient Titles => this.titles;
 
         /// <inheritdoc />
         public virtual ITokenInfoClient TokenInfo => this.tokenInfo;