+++ Gw2Sharp/WebApi/V2/Models/Skills/SkillFactStunBreak.cs
@@ -1,0 +1,13 @@
+namespace Gw2Sharp.WebApi.V2.Models
+{
+    /// <summary>
+    /// A stun break skill fact.
+    /// </summary>
+    public class SkillFactStunBreak : SkillFact
+    {
+        /// <summary>
+        /// The skill fact value.
+        /// </summary>
+        public bool Value { get; set; }
+    }
+}