+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `/v2/specializations`
 - Add `/v2/stories`
 - Add `/v2/stories/seasons`
+- Add `/v2/titles`
 
 ### Refactoring
 - **Breaking:** `Gw2Sharp.Models.MountSkinDyeSlot` has been renamed to `Gw2Sharp.Models.SkinDyeSlot`